refactor: code review cleanup for voom verify
- Fix UTF-8 panic in quick.rs::truncate
- Use file_by_path instead of path_prefix LIKE scan
- Drop redundant quarantine_dir.exists() pre-check
- Use DateTime::UNIX_EPOCH constant
- set_file_status takes FileStatus not raw string
- Drop leaked task-reference comment
